Refactor BankAccount and related classes to use BigDecimal for balance and amount; update application properties for database connection

+++ demo/src/main/java/com/bankApplication/demo/dto/WithdrawRequest.java
@@ -4,10 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class WithdrawRequest {
     private Integer id;
-    private double amount;
+    private BigDecimal amount;
 }
