+++ demo/src/main/java/com/bankApplication/demo/service/TransactionService.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Service
@@ -26,7 +27,7 @@ public class TransactionService {
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
     public Transaction createTransaction(BankAccount account,
-                                         double amount){
+                                         BigDecimal amount){
 
         Transaction transaction = new Transaction(account,amount);
